add entity mutations, queries and types
