--- frontend/src/APIClients/EntityAPIClient.ts
@@ -1,121 +1,0 @@
-import baseAPIClient from "./BaseAPIClient";
-import AUTHENTICATED_USER_KEY from "../constants/AuthConstants";
-import { getLocalStorageObjProperty } from "../utils/LocalStorageUtils";
-
-enum EnumField {
-  "A",
-  "B",
-  "C",
-  "D",
-}
-
-export type EntityRequest = {
-  stringField: string;
-  intField: number;
-  stringArrayField: string[];
-  enumField: EnumField;
-  boolField: boolean;
-};
-
-export type EntityResponse = {
-  id: string | number;
-  stringField: string;
-  intField: number;
-  stringArrayField: string[];
-  enumField: EnumField;
-  boolField: boolean;
-  fileName: string;
-};
-
-const create = async ({
-  formData,
-}: {
-  formData: FormData;
-}): Promise<EntityResponse> => {
-  const bearerToken = `Bearer ${getLocalStorageObjProperty(
-    AUTHENTICATED_USER_KEY,
-    "accessToken",
-  )}`;
-  try {
-    const { data } = await baseAPIClient.post("/entities", formData, {
-      headers: { Authorization: bearerToken },
-    });
-    return data;
-  } catch (error) {
-    return error;
-  }
-};
-
-const get = async (): Promise<EntityResponse[]> => {
-  const bearerToken = `Bearer ${getLocalStorageObjProperty(
-    AUTHENTICATED_USER_KEY,
-    "accessToken",
-  )}`;
-  try {
-    const { data } = await baseAPIClient.get("/entities", {
-      headers: { Authorization: bearerToken },
-    });
-    return data;
-  } catch (error) {
-    return error;
-  }
-};
-
-const getFile = async (uuid: string): Promise<string> => {
-  const bearerToken = `Bearer ${getLocalStorageObjProperty(
-    AUTHENTICATED_USER_KEY,
-    "accessToken",
-  )}`;
-  try {
-    const { data } = await baseAPIClient.get(`/entities/files/${uuid}`, {
-      headers: { Authorization: bearerToken },
-    });
-
-    return data.fileURL;
-  } catch (error) {
-    return error;
-  }
-};
-
-const getCSV = async (): Promise<string> => {
-  const bearerToken = `Bearer ${getLocalStorageObjProperty(
-    AUTHENTICATED_USER_KEY,
-    "accessToken",
-  )}`;
-  try {
-    const { data } = await baseAPIClient.get("/entities", {
-      // Following line is necessary to set the Content-Type header
-      // Reference: https://github.com/axios/axios/issues/86
-      data: null,
-      headers: { Authorization: bearerToken, "Content-Type": "text/csv" },
-    });
-
-    return data;
-  } catch (error) {
-    return error;
-  }
-};
-
-const update = async (
-  id: number | string,
-  {
-    entityData,
-  }: {
-    entityData: FormData;
-  },
-): Promise<EntityResponse> => {
-  const bearerToken = `Bearer ${getLocalStorageObjProperty(
-    AUTHENTICATED_USER_KEY,
-    "accessToken",
-  )}`;
-  try {
-    const { data } = await baseAPIClient.put(`/entities/${id}`, entityData, {
-      headers: { Authorization: bearerToken },
-    });
-    return data;
-  } catch (error) {
-    return error;
-  }
-};
-
-export default { create, get, getFile, getCSV, update };
