+++ frontend/src/APIClients/types/EntityClientTypes.ts
@@ -1,0 +1,25 @@
+
+enum EnumField {
+    "A",
+    "B",
+    "C",
+    "D",
+}
+
+export type EntityRequest = {
+    stringField: string;
+    intField: number;
+    stringArrayField: string[];
+    enumField: EnumField;
+    boolField: boolean;
+  };
+  
+  export type EntityResponse = {
+    id: string | number;
+    stringField: string;
+    intField: number;
+    stringArrayField: string[];
+    enumField: EnumField;
+    boolField: boolean;
+    fileName: string;
+  };
